Fix excel template

+++ eFormAPI/Plugins/MachineArea.Pn/MachineArea.Pn/Infrastructure/Models/Report/ReportEntityModel.cs
@@ -4,7 +4,7 @@ namespace MachineArea.Pn.Infrastructure.Models.Report
 {
     public class ReportEntityModel
     {
-        public int TotalTime { get; set; }
+        public decimal TotalTime { get; set; }
         public List<decimal> TimePerTimeUnit { get; set; }
         public string EntityName { get; set; }
         public int EntityId { get; set; }
